Create Contributing.md
Hoping these are the necessary steps to contribute. I'd like to contribute but the instructions are missing. The project won't load in my VS 2013 because it targets .NET Framework 4.6 so I've guessed VS2015 is required.

Please review the steps to contribute and alter them if you wish so that people like me can contribute.